separate isAppActive hook
